refactored to use double precision

+++ rayTracer/src/camera.cpp
@@ -1,7 +1,7 @@
 #include "camera.h"
 
-Camera::Camera(glm::vec4 position, glm::vec4 centre, glm::vec4 up, int hsize,
-               int vsize, float fov) : Shape()
+Camera::Camera(glm::dvec4 position, glm::dvec4 centre, glm::dvec4 up, int hsize,
+               int vsize, double fov) : Shape()
 {
   this->position = position;
   this->centre = centre;
@@ -15,16 +15,16 @@ Camera::Camera(glm::vec4 position, glm::vec4 centre, glm::vec4 up, int hsize,
 }
 
 void Camera::updateTransform() {
-  this->transform = glm::lookAt(glm::vec3(position),
-                                glm::vec3(this->centre), glm::vec3(this->up));
+  this->transform = glm::lookAt(glm::dvec3(position),
+                                glm::dvec3(this->centre), glm::dvec3(this->up));
 }
 
 Camera::~Camera() {}
 
 void Camera::setPixelSize()
 {
-  float halfView = glm::tan(this->fov / 2);
-  float aspect = (float)this->hsize / (float)this->vsize;
+  double halfView = glm::tan(this->fov / 2);
+  double aspect = (double)this->hsize / (double)this->vsize;
 
   if (aspect >= 1)
   {
@@ -36,34 +36,34 @@ void Camera::setPixelSize()
     this->halfWidth = halfView * aspect;
     this->halfHeight = halfView;
   }
-  this->pixelSize = (this->halfWidth * 2.f) / (float)this->hsize;
+  this->pixelSize = (this->halfWidth * 2.0) / (double)this->hsize;
 }
 
-Ray Camera::rayForPixel(float px, float py, int currentRayNumber, int sqrtRaysPerPixel, float halfSubPixelSize)
+Ray Camera::rayForPixel(double px, double py, int currentRayNumber, int sqrtRaysPerPixel, double halfSubPixelSize)
 {
   int subPixelRowNumber = currentRayNumber / sqrtRaysPerPixel;
   int subPixelColNumber = currentRayNumber % sqrtRaysPerPixel;
-  float subPixelxOffset = halfSubPixelSize * subPixelColNumber;
-  float subPixelyOffset = halfSubPixelSize * subPixelRowNumber;
+  double subPixelxOffset = halfSubPixelSize * subPixelColNumber;
+  double subPixelyOffset = halfSubPixelSize * subPixelRowNumber;
 
-  float xOffset = (px + subPixelxOffset) * this->pixelSize;
-  float yOffset = (py + subPixelyOffset) * this->pixelSize;
+  double xOffset = (px + subPixelxOffset) * this->pixelSize;
+  double yOffset = (py + subPixelyOffset) * this->pixelSize;
 
-  float worldX = this->halfWidth - xOffset;
-  float worldY = this->halfHeight - yOffset;
+  double worldX = this->halfWidth - xOffset;
+  double worldY = this->halfHeight - yOffset;
 
-  glm::vec4 pixel = glm::affineInverse(this->transform) *
-                    glm::vec4(worldX, worldY, -1.f, 1.f);
-  glm::vec4 origin =
-      glm::affineInverse(this->transform) * glm::vec4(0.f, 0.f, 0.f, 1.f);
-  glm::vec4 direction = glm::normalize(pixel - origin);
+  glm::dvec4 pixel = glm::affineInverse(this->transform) *
+                    glm::dvec4(worldX, worldY, -1.0, 1.0);
+  glm::dvec4 origin =
+      glm::affineInverse(this->transform) * glm::dvec4(0.0, 0.0, 0.0, 1.0);
+  glm::dvec4 direction = glm::normalize(pixel - origin);
 
   return Ray(origin, direction);
 }
 
-glm::vec4 Camera::normalAt(glm::vec4 point)
+glm::dvec4 Camera::normalAt(glm::dvec4 point)
 {
-  return glm::vec4();
+  return glm::dvec4();
 }
 
 void Camera::intersectRay(Ray& ray, std::vector<Geometry::Intersection<Shape>>& intersections) { return; }
