+++ rayTracer/src/cylinder.h
@@ -8,19 +8,19 @@
 class Cylinder : public Shape {
 public:
   Cylinder();
-  Cylinder(float minimum, float maximum, bool capped);
+  Cylinder(double minimum, double maximum, bool capped);
   ~Cylinder();
 
   virtual void intersectRay(Ray& ray, std::vector<Geometry::Intersection<Shape>>& intersections) override;
-  virtual glm::vec4 normalAt(glm::vec4 point) override;
+  virtual glm::dvec4 normalAt(glm::dvec4 point) override;
   virtual std::string type() override;
 
-  float minimum;
-  float maximum;
+  double minimum;
+  double maximum;
   bool capped;
 
 private:
-  bool checkCap(Ray &ray, float t);
+  bool checkCap(Ray &ray, double t);
 
   void intersectCaps(Ray &ray, std::vector<Geometry::Intersection<Shape>>& intersections);
 };
