+++ rayTracer/src/plane.cpp
@@ -14,19 +14,19 @@ void Plane::intersectRay(Ray& ray, std::vector<Geometry::Intersection<Shape>>& i
     if (std::abs(transformedRay.direction.y) < Geometry::EPSILON)
         return;
 
-    float t = -transformedRay.origin.y / transformedRay.direction.y;
+    double t = -transformedRay.origin.y / transformedRay.direction.y;
 
     intersections.push_back(Geometry::Intersection<Shape>{t, this});
 
 }
 
-glm::vec4 Plane::normalAt(glm::vec4 point)
+glm::dvec4 Plane::normalAt(glm::dvec4 point)
 {
-    // glm::mat4 transformInverse(glm::affineInverse(this->transform));
-    // glm::vec4 objectPoint = transformInverse * point;
-    glm::vec4 objectNormal(0.f, 1.f, 0.f, 0.f);
-    glm::vec4 worldNormal = glm::transpose(this->inverseTransform) * objectNormal;
-    worldNormal.w = 0.f;
+    // glm::dmat4 transformInverse(glm::affineInverse(this->transform));
+    // glm::dvec4 objectPoint = transformInverse * point;
+    glm::dvec4 objectNormal(0.0, 1.0, 0.0, 0.0);
+    glm::dvec4 worldNormal = glm::transpose(this->inverseTransform) * objectNormal;
+    worldNormal.w = 0.0;
 
     return glm::normalize(worldNormal);
 }
