+++ rayTracer/src/pointLight.cpp
@@ -1,13 +1,13 @@
 #include "pointLight.h"
 
-PointLight::PointLight(glm::vec4 position, glm::vec3 intensity) : Shape() {
+PointLight::PointLight(glm::dvec4 position, glm::dvec3 intensity) : Shape() {
   this->position = position;
   this->intensity = intensity;
 }
 
 PointLight::~PointLight() {}
 
-glm::vec4 PointLight::normalAt(glm::vec4 point) { return glm::vec4(); }
+glm::dvec4 PointLight::normalAt(glm::dvec4 point) { return glm::dvec4(); }
 
 void PointLight::intersectRay(Ray& ray, std::vector<Geometry::Intersection<Shape>>& intersections) { return; }
